+++ src/menelaus/ensemble/__init__.py
@@ -1,0 +1,5 @@
+from menelaus.ensemble.election import SimpleMajorityElection, MinimumApprovalElection
+from menelaus.ensemble.election import OrderedApprovalElection, ConfirmedElection
+
+from menelaus.ensemble.ensemble import Ensemble
+from menelaus.ensemble.ensemble import BatchEnsemble, StreamingEnsemble
